+++ features/review/index.ts
@@ -1,5 +1,4 @@
 export { ReviewSubmitDialog } from './components/review-submit-dialog';
-export { ReviewWriteForm } from './components/review-write-form';
 export { WriteReviewFlow } from './components/write-review/write-review-flow';
 export { UserReviewCardLoading } from './components/user-review-card-loading';
 export { fetchReviewFormData, fetchReviewValidation, reviewKeys } from './review.api';
