+++ mocks/handlers/users.ts
@@ -30,7 +30,7 @@ const mockUserReviews = [
     reviewerNickname: '김소스',
     feedback:
       '코드 리뷰를 꼼꼼하게 해주시고, 모르는 부분도 편하게 질문할 수 있는 분위기를 만들어 주셔서 협업이 수월했어요. 문서 정리도 깔끔해서 온보딩이 빨랐습니다.',
-    projectPosition: 'FRONTEND',
+    projectPosition: 'FE',
     projectStatus: 'COMPLETED',
   },
   {
@@ -42,7 +42,7 @@ const mockUserReviews = [
     reviewerNickname: '박바',
     feedback:
       '피드백을 빠르게 반영해 주시고, 팀원들 의견을 잘 조율해 주셨어요. 덕분에 마감 전에 퀄리티 있는 결과물을 낼 수 있었습니다.',
-    projectPosition: 'BACKEND',
+    projectPosition: 'BE',
     projectStatus: 'COMPLETED',
   },
 ] as const;
